Update bash files

+++ analysis/shellcast-analysis/src/fl_pqpf/xmrg_proc.sh
@@ -1,13 +1,12 @@
 #!/bin/sh
-CUR_DIR="${PWD%/*}"
-ROOT="$(dirname "$CUR_DIR")"
-echo $ROOT
-RAW_DATA_DIR=$ROOT'/data/tp/raw'
-INTERMEDIATE_DIR=$ROOT'/data/tp/fl/intermediate'
-TP_OUTPUTS_DIR=$ROOT'/data/tp/fl/outputs'
+
+echo "Starting xmrg_proc.sh"
+RAW_DATA_DIR=$PWD'/data/tp/raw'
+INTERMEDIATE_DIR=$PWD'/data/tp/fl/intermediate'
+TP_OUTPUTS_DIR=$PWD'/data/tp/fl/outputs'
 INTERMEDIATE_PROC_DIR=$INTERMEDIATE_DIR'/procs'
 CNV_GRIB2_DIR=$INTERMEDIATE_DIR'/cnvgrib2'
-NCEPLIB=$ROOT'/ncep-lib-utils/nceplibs/bin'
+NCEPLIB=$PWD'/ncep-lib-utils/nceplibs/bin'
 
 rm -rf $CNV_GRIB2_DIR
 mkdir $CNV_GRIB2_DIR
@@ -15,12 +14,22 @@ mkdir $CNV_GRIB2_DIR
 # Convert GRIB1  to GRIB2
 counter=0
 raw_gribs=`ls $RAW_DATA_DIR/*.grb`
+count_gribs=`ls $RAW_DATA_DIR | wc -l`
+echo "Number of GRIB1 files: $count_gribs"
 for grib in $raw_gribs
 do
   fname=`basename $grib`
   `$NCEPLIB/cnvgrib -g12 $grib $CNV_GRIB2_DIR/$fname`
   counter=$((counter+1))
 done
+
+cnv_gribs=`ls $CNV_GRIB2_DIR/*.grb | wc -l`
+echo "Number of GRIB2 files: $cnv_gribs"
+if [ $cnv_gribs -ne $count_gribs ]
+then
+  echo "Error: Not all GRIB1 files were converted to GRIB2"
+  exit 1
+fi
 
 # ----- Clean directory -----
 rm -f -R $TP_OUTPUTS_DIR
@@ -56,3 +65,7 @@ for ((i=24; i<=$counter; i+=24 ));
     gdalwarp -of GTiff -t_srs EPSG:4326 'tp_'$i'h'.nc $outfile
 done
 
+tiffs=`ls $TP_OUTPUTS_DIR/*.tif | wc -l`
+echo "Number of TIFF files: $tiffs"
+
+echo "Finished xmrg_proc.sh"
